rename root_finding_1_var to nonlinear_eqn
