+++ csnum/nonlinear_eqn.py
